requirements.txt
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.30.0
pycocotools
tqdm
pyyaml
Pillow
scipy
timm
matplotlib
numpy
seaborn
pandas
scikit-learn
opencv-python
scipy
torchmetrics
albumentations